+++ discord/objects/__init__.py
@@ -10,6 +10,7 @@ from .auto_moderation_rule_trigger_metadata import AutoModerationRuleTriggerMeta
 from .avatar_decoration_data import AvatarDecorationData
 from .channel import Channel
 from .channel_info_channel import ChannelInfoChannel
+from .client_status import ClientStatus
 from .collectible import Collectible
 from .default_reaction import DefaultReaction
 from .emoji import Emoji