+++ discord/objects/thread_member.py
@@ -1,4 +1,4 @@
-# from ..events import PresenceUpdateEvent
+from ..events import PresenceUpdateEvent
 from .._dataclass import dataclass
 from ..snowflake import Snowflake
 from ..utils import ISO8601Timestamp, Nullable, Optional
@@ -29,7 +29,7 @@ class ThreadMember:
       Only present when ``with_member`` is set to ``True`` when calling :meth:`~discord._http.HTTP.list_thread_members` or :meth:`~discord._http.HTTP.get_thread_member`.
   """
 
-  presence: Optional[Nullable["PresenceUpdateEvent"]]
+  presence: Optional[Nullable[PresenceUpdateEvent]]
 
   user_id: Optional[Snowflake]
   """ID of the user.